Simplify how _AggregateMatrices is tested
