+++ tests/unit/autojac/_transform/utils/__init__.py
@@ -1,3 +1,1 @@
 from .dict_assertions import assert_tensor_dicts_are_close
-from .inputs import aggregator, jacobian_matrix_dicts, keys
-from .property_testers import EmptyDictProperty, ExpectedStructureProperty
